% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_volumes.R
\name{mesh_volumes}
\alias{mesh_volumes}
\title{mesh_volumes}
\usage{
mesh_volumes(surfaces = surfaces)
}
\arguments{
\item{surfaces}{an object of the class "mesh_array" that contains two lists: a table with the meshes information (name, number of faces and vertices) and all the meshes that had previously been imported using the function import_surfaces.}
}
\value{
This function returns a table with the meshes information (name, number of faces and vertices) and the volume for each surface mesh.
}
\description{
Estimates the volume of multiple meshes' surfaces.
}
\details{
Takes an object of the class "mesh_array" and returns the volume for each mesh.
  For more information check the R package Rvcg (Schlager, 2017)
}
\examples{
data(meshes)
vol <- mesh_volumes(meshes)

}
\references{
Schlager S (2017). “Morpho and Rvcg - Shape Analysis in R.” In Zheng G, Li S, Szekely G (eds.), _Statistical Shape and Deformation Analysis_, 217-256. Academic Press. ISBN 9780128104934.
}
\author{
Marta Vidal-Garcia
}
