% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mps2array.R
\name{mps2array}
\alias{mps2array}
\title{mps2array}
\usage{
mps2array(dir = NULL, ID = NULL, save.txt = FALSE)
}
\arguments{
\item{dir}{optional variable that selects which directory the mps files
should be imported from. The default is the current directory}

\item{ID}{optional vector with IDs for each landmark matrix from the mps files}

\item{save.txt}{optional. It saves each of the 3D landmark matrices as txt files,
without row or column names.}
}
\description{
Imports *.mps files into an array of 3D landmarks.
}
\details{
Takes multiple mps files and batch-imports them as a landmark. It can also
  save the landmarks for each specimen as txt files.
}
\examples{

dir(pattern = "*.mps") # three mps files with five 3D landmarks each
land_array <- mps2array(dir = "./data")

# Remove "skull" from the ID name, so that the IDs are "spec1", "spec2", and "spec3".
land_array_2 <- mps2array(dir = "./data", save.txt = TRUE)

# Assign IDs for each landmark matrix.
ID_land <- c("specimen1", "specimen2", "specimen3")
land_array_3 <- mps2array(dir = "./data", ID = ID_land)

}
\author{
Marta Vidal-Garcia
}
