% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ply2vtk.R
\name{ply2vtk}
\alias{ply2vtk}
\title{ply2vtk}
\usage{
plyvtk(files = NULL, ID = NULL, vtk_path = NULL, print.progress = TRUE)
}
\arguments{
\item{files}{Optional parameter indicating where to save the VTK files. The default is the current directory.}

\item{ID}{optional vector with IDs used for saving the VTK meshes. Default is the same name as the PLY meshes}

\item{print.progress}{optional parameter indicating whether to print the progress}
}
\value{
This function returns a table with the meshes information (name, number of faces and vertices) and the volume for each surface mesh.
}
\description{
Batch-converts PLY mesh files into VTK mesh files
}
\details{
Takes a a list of PLY meshes and transforms those files into VTK. For more information check the R package _freesurferformats_ (Schäfer, 2020)
}
\examples{
# Convert all PLY meshes to VTK meshes
setwd("./data")
dir(pattern = "*.ply")
ply2vtk(files = NULL, ID = NULL, vtk_path = NULL, print.progress = TRUE)
setwd("../")

# Convert a subset PLY meshes to VTK meshes, and change ID
setwd("./data")
files_ply <- dir(pattern = "*.ply")[1:2]
ID_vtk <- c("specimen1", "specimen2")
ply2vtk(files = files, ID = NULL, vtk_path = NULL, print.progress = TRUE)
setwd("../")

}
\references{
Schlager S (2017). “Morpho and Rvcg - Shape Analysis in R.” In Zheng G, Li S, Szekely G (eds.), _Statistical Shape and Deformation Analysis_, 217-256. Academic Press. ISBN 9780128104934.
Tim Schäfer (2020). freesurferformats: Read and Write 'FreeSurfer' Neuroimaging File Formats. R package version 0.1.10. https://CRAN.R-project.org/package=freesurferformats
}
\author{
Marta Vidal-Garcia
}
