% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_surfaces.R
\name{read_surfaces}
\alias{read_surfaces}
\title{read_surfaces}
\usage{
read_surfaces <- function(dir = NULL, pattern = pattern, files = NULL,
  target_faces = NULL, dec_percent = NULL, mesh_topo = FALSE, mesh_qual = FALSE,
  mesh_bound = FALSE, mergeVert = FALSE, vcgIsolated = FALSE, IsoFacenum = NULL,
  IsoDiameter = NULL, clean = FALSE, cleansed = NULL)
}
\arguments{
\item{pattern}{string parameter that determines what type of mesh to import (e.g. 'ply' or 'stl')}

\item{files}{Optional parameter. The default is all the files with a pattern (e.g. ply or stl).}

\item{target_faces}{Number of target faces for all the meshes. For more information check *?vcgQEdecim* in Rvcg.}

\item{dec_percent}{Degree of face reduction (from 0 to 1) across all meshes. For more information check *?vcgQEdecim* in Rvcg.}

\item{mesh_topo}{Optional logical parameter, indicating if the mesh topology will be preserved across all the meshes. The default is FALSE. For more information check *?vcgQEdecim* in Rvcg.}

\item{mesh_qual}{Optional logical parameter, indicating whether vertex quality will be considered across all the meshes. The default is FALSE. For more information check *?vcgQEdecim* in Rvcg.}

\item{mesh_bound}{Optional logical parameter, indicating whether mesh boundary will be preserved across all the meshes. The default is FALSE. For more information check *?vcgQEdecim* in Rvcg.}

\item{mergeVert}{Optional logical parameter to merge duplicated vertices across all meshes. For more information check *?mergeVertices* in rgl.}

\item{vcgIsolated}{Optional parameter to remove isolated pieces. The default is FALSE. For more information check *?vcgIsolated* in Rvcg.}

\item{IsoFacenum}{Optional parameter that determines the threshold to remove isolated pieces. The default is FALSE. For more information check *?vcgIsolated* in Rvcg.}

\item{IsoDiameter}{Optional parameter that determines the threshold to remove isolated pieces (pices with an smaller diamter are removed). The default is FALSE. For more information check *?vcgIsolated* in Rvcg.}

\item{clean}{Optional cleaning parameter from Rvcg that batch-cleans all meshes. The default is FALSE. For more information check *?vcgIsolated* in Rvcg.}

\item{dir}{Optional parameter if the directory that contains the mesh is other than the current working directory. The default (NULL) is the current working directory.}

\item{cleansel}{Optional cleaning parameter from Rvcg that determines which cleanign algorithms are applied across all meshes and in which order. The default (NULL) is options 0 to 6. For more information check *?vcgIsolated* in Rvcg.}
}
\value{
This function returns an object of the class 'mesh_array' will keep the dimnames
  associated with each mesh (which correspond to the filenames).
}
\description{
Imports multiple meshes' surfaces.
}
\details{
Imports multiple mesh files (e.g. PLY, STL) and batch-processes them (e.g. reducing the number of faces, merging vertices, removing isolated pieces, etc. with a few functions from Rvcg). For more information check the R package Rvcg (Schlager, 2017)
}
\examples{
# PLY meshes
dir <- "./data/"

# Import PLY meshes without any processing
test1 <- read_surfaces(dir = "./data", pattern = "ply")

# Import STL meshes and reduce number of faces
test2 <- read_surfaces(dir = dir, files = NULL, pattern = "stl", target_faces = 50000, clean = TRUE)

}
\references{
Schlager S (2017). “Morpho and Rvcg - Shape Analysis in R.” In Zheng G, Li S, Szekely G (eds.), _Statistical Shape and Deformation Analysis_, 217-256. Academic Press. ISBN 9780128104934.
}
\author{
Marta Vidal-Garcia
}
