% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mps.R, R/write_surfaces.R
\name{write_mps}
\alias{write_mps}
\alias{write_surfaces}
\title{write_mps}
\usage{
write_mps(arrayLMs = arrayLMs, dir = NULL, ID = NULL)

write_surfaces(meshes = meshes, dir = NULL, pattern = c("PLY", "STL", "VTK"),
  binary = FALSE, print.progress = TRUE)
}
\arguments{
\item{arrayLMs}{an object of the class "array" that contains
three-dimensional landmarks for 'n' specimens}

\item{dir}{optional variable that selects which directory the mps files should be saved to}

\item{ID}{optional vector with IDs for saving the *.mps files}

\item{meshes}{XXX}

\item{pattern}{optional vector with IDs for saving the *.mps files}

\item{binary}{optional variable that selects which directory the mps files should be saved to}

\item{print.progress}{optional vector with IDs for saving the *.mps files}
}
\description{
Writes *.mps files from an array of 3D landmarks.

Writes meshes as PLY, STL, or VTK files
}
\details{
Takes a landmark array and saves the landmarks for each specimen as *.mps files.

Takes a landmark array and saves the landmarks for each specimen as *.mps files.
}
\examples{
test1 <- array(data = 1:10, dim = c(5,3,2))
write_mps(arrayLMs = test1, ID = c("spec1", "spec2"))

test2 <- array(data = 1:10, dim = c(5,3,2), dimnames = list(c("LM1", "LM2",
"LM3", "LM4", "LM5"), c("x", "y", "z"), c("spec1", "spec2")))
write_mps(arrayLMs = test2)

test3 <- array(data = 1:10, dim = c(5,3,2))
dimnames(test3)[[3]] <- c("spec1", "spec2")
write_mps(arrayLMs = test3, dir = "../")

test1 <- array(data = 1:10, dim = c(5,3,2))
write_mps(arrayLMs = test1, ID = c("spec1", "spec2"))

test2 <- array(data = 1:10, dim = c(5,3,2), dimnames = list(c("LM1", "LM2",
"LM3", "LM4", "LM5"), c("x", "y", "z"), c("spec1", "spec2")))
write_mps(arrayLMs = test2)

test3 <- array(data = 1:10, dim = c(5,3,2))
dimnames(test3)[[3]] <- c("spec1", "spec2")
write_mps(arrayLMs = test3, dir = "../")

}
\author{
Marta Vidal-Garcia

Marta Vidal-Garcia
}
