% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mps.R
\name{write_mps}
\alias{write_mps}
\title{write_mps}
\usage{
write_mps(arrayLMs = arrayLMs, dir = NULL, ID = NULL)
}
\arguments{
\item{arrayLMs}{an object of the class "array" that contains
three-dimensional landmarks for 'n' specimens}

\item{dir}{optional variable that selects which directory the mps files should be saved to}

\item{ID}{optional vector with IDs for saving the *.mps files}
}
\description{
Writes *.mps files from an array of 3D landmarks.
}
\details{
Takes a landmark array and saves the landmarks for each specimen as *.mps files.
}
\examples{
test1 <- array(data = 1:10, dim = c(5,3,2))
write_mps(arrayLMs = test1, ID = c("spec1", "spec2"))

test2 <- array(data = 1:10, dim = c(5,3,2), dimnames = list(c("LM1", "LM2",
"LM3", "LM4", "LM5"), c("x", "y", "z"), c("spec1", "spec2")))
write_mps(arrayLMs = test2)

test3 <- array(data = 1:10, dim = c(5,3,2))
dimnames(test3)[[3]] <- c("spec1", "spec2")
write_mps(arrayLMs = test3, dir = "../")

}
\author{
Marta Vidal-Garcia
}
